requirements.txt
Django>=2.2.10,<2.3
django-ckeditor>=5.9.0,<5.10
django-photologue>=3.11,<3.12
django-reversion>=3.0.5,<3.1
django-mapbox-location-field>=1.4.0,<1.5
django-mptt>=0.11.0,<0.12
